requirements.txt
streamlit
streamlit-authenticator
requests
python-dotenv
flask
flask_sqlalchemy
authlib
pyjwt
flask-cors 